+++ src/adapters/http/routes/balances.rs
@@ -986,6 +986,8 @@ mod tests {
             version: env!("CARGO_PKG_VERSION"),
             database_pool: None,
             api_key_repository: None,
+            api_key_minute_limiter: crate::adapters::http::rate_limit::ApiKeyMinuteLimiter::default(
+            ),
             asset_repository: Some(GlobalAssetRepository::in_memory(sample_assets())),
             price_indexer_client,
             dis_client: None,